Taking care of test flight configuration and small warnings fixes

+++ zither/Classes/EditProduct/SSIWarrantyPicker.m
@@ -151,7 +151,7 @@
 {
     if (component == 0) {
 
-        return [NSString stringWithFormat:@"%lu", row];
+        return [NSString stringWithFormat:@"%lu", (long)row];
     }
     else {
